Fix parsing of `adb shell ps' output
The regex to parse the output lines from `adb shell ps' expected the
`WCHAN' field to have a numeric value. This results in an empty list of
PIDs to filter for, which in turn results in no output, since to log
line matches.

Update the regex to expect an alphanumeric value in the `WCHAN' field.

Github: Fixes #4
Signed-off-by: Florian Limberger <flo@snakeoilproductions.net>