Implement filters as goroutines
To simplify the code a bit, implement filters as goroutines which are
interconnected by channels. Instead of returning a boolean for lines to
be allowed to be processed, write lines to the output channel. Filtered
lines are not written and therefore swallowed by the filter function.

This also simplifies the output handling, which is also implemented as a
goroutine.

Signed-off-by: Florian Limberger <flo@purplekraken.com>