Fix binary output in file redirection
On terminals, colours are rendered by outputting escape sequences. These
are interpreted as binary when encountered in files, and it is customary
for many command-line utilities to check whether the output goes to a
terminal before inserting those escape sequences into the output. Fis
this issue by adopting that behaviour and checking whether `os.Stdout`
is a TTY. Initialize Go modules to manage the external dependency
`github.com/mattn/go-isatty`.

This introduces the first external dependency, which is a bit sad for me.

Signed-off-by: Florian Limberger <flo@snakeoilproductions.net>
Closes: #8